# docs/api/openapi-v1.yaml
openapi: 3.0.3
info:
  title: API de Integração Financeira ACOP
  version: 1.1.0
  description: |
    API servidor-servidor para entrada e consulta de Payables, Receivables e fatos bancários.
    O Bearer token determina o sistema de origem. Toda mutação exige
    `Idempotency-Key`; `X-Correlation-ID` é opcional e sempre é devolvido.
servers:
  - url: http://localhost/api/v1
    description: Ambiente local; produção deve usar HTTPS
security:
  - bearerAuth: []
...
paths:
  /payables:
    post:
      tags: [Payables]
      summary: Criar ou reingerir uma conta a pagar
      description: Exige o scope `payables:write`. A rota impõe o tipo `PAYABLE`.
      operationId: createPayable
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialTitleCreate'
            example:
              external_id: PAG-2026-000456
              document_number: FAT-9988
              issue_date: '2026-08-13'
              due_date: '2026-08-20'
              original_amount: '830.50'
              discount_amount: '0.00'
              addition_amount: '0.00'
              currency: BRL
              party:
                type: SUPPLIER
                name: Fornecedor Exemplo
              installment_count: 1
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /payables/{external_id}:
    parameters:
      - $ref: '#/components/parameters/ExternalId'
    get:
      tags: [Payables]
      summary: Consultar conta a pagar pela identidade externa
      description: Exige `payables:read`; restringe origem autenticada e tipo PAYABLE.
      operationId: getPayable
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Payables]
      summary: Atualizar uma conta a pagar editável
      description: Exige `payables:write`. O external_id vem do path e não pode estar no corpo.
      operationId: updatePayable
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialTitleUpdate'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /receivables:
    post:
      tags: [Receivables]
      summary: Criar ou reingerir uma conta a receber
      description: Exige o scope `receivables:write`. A rota impõe o tipo `RECEIVABLE`.
      operationId: createReceivable
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialTitleCreate'
            example:
              external_id: REC-2026-000123
              document_number: NF-12345
              issue_date: '2026-08-13'
              due_date: '2026-09-13'
              original_amount: '1500.00'
              discount_amount: '0.00'
              addition_amount: '0.00'
              currency: BRL
              party:
                type: CUSTOMER
                name: Cliente Exemplo
              installment_count: 1
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /receivables/{external_id}:
    parameters:
      - $ref: '#/components/parameters/ExternalId'
    get:
      tags: [Receivables]
      summary: Consultar conta a receber pela identidade externa
      description: Exige `receivables:read`; restringe origem autenticada e tipo RECEIVABLE.
      operationId: getReceivable
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags: [Receivables]
      summary: Atualizar uma conta a receber editável
      description: Exige `receivables:write`. O external_id vem do path e não pode estar no corpo.
      operationId: updateReceivable
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinancialTitleUpdate'
      responses:
        '200':
          $ref: '#/components/responses/Success'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token opaco `acop_...`; scopes pertencem à credencial cadastrada.
      x-scopes:
        - payables:read
        - payables:write
        - receivables:read
        - receivables:write
        - bank-transactions:read
        - bank-transactions:write
        - bank-imports:read
        - bank-imports:write
  parameters:
    ExternalId:
      name: external_id
      in: path
      required: true
      schema:
        type: string
        maxLength: 128
      description: Identificador dentro da origem autenticada.
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
      description: Chave de um evento lógico; repetir exatamente para recuperar a resposta.
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: Correlação opcional; UUID é gerado quando ausente.
  headers:
    CorrelationId:
      description: Correlação efetiva da operação.
      schema:
        type: string
    IdempotencyReplayed:
      description: "`true` quando a resposta veio da inbox."
      schema:
        type: string
        enum: ['true', 'false']
  schemas:
    Money:
      type: string
      pattern: '^(0|[1-9][0-9]{0,12})\.[0-9]{2}$'
      example: '1500.00'
    Party:
      type: object
      additionalProperties: false
      properties:
        id:
          type: integer
          minimum: 1
          nullable: true
        type:
          type: string
          maxLength: 30
          nullable: true
        name:
          type: string
          maxLength: 191
          nullable: true
    FinancialTitleFields:
      type: object
      additionalProperties: false
      required: [issue_date, due_date, original_amount]
      properties:
        external_id:
          type: string
          maxLength: 128
        document_number:
          type: string
          maxLength: 120
          nullable: true
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        original_amount:
          $ref: '#/components/schemas/Money'
        discount_amount:
          allOf:
            - $ref: '#/components/schemas/Money'
          default: '0.00'
        addition_amount:
          allOf:
            - $ref: '#/components/schemas/Money'
          default: '0.00'
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          default: BRL
        party:
          $ref: '#/components/schemas/Party'
        account_id:
          type: integer
          minimum: 1
          nullable: true
        category_id:
          type: integer
          minimum: 1
          nullable: true
        cost_center_id:
          type: integer
          minimum: 1
          nullable: true
        installment_count:
          type: integer
          minimum: 1
          maximum: 999
          default: 1
        notes:
          type: string
          maxLength: 10000
          nullable: true
    FinancialTitleCreate:
      allOf:
        - $ref: '#/components/schemas/FinancialTitleFields'
        - type: object
          required: [external_id]
          properties:
            external_id:
              type: string
              maxLength: 128
    FinancialTitleUpdate:
      allOf:
        - $ref: '#/components/schemas/FinancialTitleFields'
        - not:
            required: [external_id]
      description: Mesmo contrato, sem external_id, type, source_system ou campos internos.
...
    Installment:
      type: object
      required: [number, due_date, amount, status]
      properties:
        number:
          type: integer
        due_date:
          type: string
          format: date
        amount:
          $ref: '#/components/schemas/Money'
        status:
          type: string
          enum: [OPEN, PARTIALLY_SETTLED, SETTLED, CANCELLED]
    FinancialTitle:
      type: object
      required: [id, external_id, type, status, issue_date, due_date, total_amount, currency]
      properties:
        id:
          type: integer
        external_id:
          type: string
        type:
          type: string
          enum: [PAYABLE, RECEIVABLE]
        status:
          type: string
          enum: [OPEN, PARTIALLY_SETTLED, SETTLED, CANCELLED]
        document_number:
          type: string
          nullable: true
        issue_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        original_amount:
          $ref: '#/components/schemas/Money'
        discount_amount:
          $ref: '#/components/schemas/Money'
        addition_amount:
          $ref: '#/components/schemas/Money'
        total_amount:
          $ref: '#/components/schemas/Money'
        currency:
          type: string
        party:
          $ref: '#/components/schemas/Party'
        installment_count:
          type: integer
        installments:
          type: array
          items:
            $ref: '#/components/schemas/Installment'
        notes:
          type: string
          nullable: true
        cancellation:
          type: object
          properties:
            reason:
              type: string
            cancelled_at:
              type: string
              format: date-time
...
    SuccessEnvelope:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/FinancialTitle'
        meta:
          type: object
          required: [correlation_id]
          properties:
            correlation_id:
              type: string
            idempotency_replayed:
              type: boolean
            decision:
              type: string
              enum: [CREATED, UPDATED, IGNORED, CANCELLED]
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details, correlation_id]
          properties:
            code:
              type: string
              enum:
                - UNAUTHENTICATED
                - FORBIDDEN
                - VALIDATION_ERROR
                - SOURCE_SYSTEM_INACTIVE
                - RESOURCE_NOT_FOUND
                - METHOD_NOT_ALLOWED
                - IDEMPOTENCY_KEY_REQUIRED
                - IDEMPOTENCY_KEY_REUSED
                - IDEMPOTENCY_REQUEST_IN_PROGRESS
                - TITLE_UPDATE_NOT_ALLOWED
                - TITLE_CANCEL_NOT_ALLOWED
                - TITLE_ALREADY_SETTLED
                - DOMAIN_CONFLICT
                - BANK_ACCOUNT_NOT_FOUND
                - BANK_TRANSACTION_ID_REQUIRED
                - BANK_TRANSACTION_ID_CONFLICT
                - BANK_IMPORT_INVALID_FILE
                - BANK_IMPORT_UNSUPPORTED_FORMAT
                - BANK_IMPORT_TOO_LARGE
                - BANK_IMPORT_FAILED
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            correlation_id:
              type: string
  responses:
    Created:
      description: Título criado.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
        Idempotency-Replayed:
          $ref: '#/components/headers/IdempotencyReplayed'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    Success:
      description: Consulta, atualização, ignorado, cancelamento ou replay concluído.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
        Idempotency-Replayed:
          $ref: '#/components/headers/IdempotencyReplayed'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SuccessEnvelope'
    BadRequest:
      $ref: '#/components/responses/ErrorBody'
    Unauthenticated:
      $ref: '#/components/responses/ErrorBody'
    Forbidden:
      $ref: '#/components/responses/ErrorBody'
    NotFound:
      $ref: '#/components/responses/ErrorBody'
    Conflict:
      $ref: '#/components/responses/ErrorBody'
    ValidationError:
      $ref: '#/components/responses/ErrorBody'
...
    RateLimited:
      $ref: '#/components/responses/ErrorBody'
    InternalError:
      $ref: '#/components/responses/ErrorBody'
    ErrorBody:
      description: Erro padronizado.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'